♻️ firebase 코드를 context에서 api 파일로 변경
